[BLUO-573] - Refatoracao dos metodos de delete no controller e service

+++ candidateCIandT/src/main/java/br/com/candidate/candidateCIandT/candidate/CandidateService.java
@@ -23,7 +23,11 @@ public class CandidateService {
         return candidateRepository.findByStatus(status);
     }
 
-    public void deleteCandidate(Long id){
-        candidateRepository.deleteById(id);
+    public boolean deleteCandidate(Long id){
+        if (candidateRepository.existsById(id)){
+            candidateRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
